Refactor: restructure project by creating dedicated subdirectories
- Improved project structure by organising files into dedicated subdirectories.
- Created new subdirectories `include/binarytree/concepts` and `include/binarytree/tree_helpers` for header files.
- Created analogous subdirectories `src/binarytree/concepts` and `src/binarytree/tree_helpers` for source (`.tpp`) files.
- Moved `Iterator` class definition to `include/binarytree/tree_helpers/tree_iterator.h`.
- Moved `Iterator` class implementation to `src/binarytree/tree_helpers/tree_iterator.tpp`.
- Ensured that the `Iterator` class works correctly with the `TreeNode<T>` structure.
- Updated `CMakelists.txt` with appropriate `target_include_directories` for the new subdirectories.
